requirements.txt
Pillow
imagehash
streamlit
